migrated from cache to newcache
