+++ procastro/astrofile/astrofile.py
@@ -13,7 +13,7 @@ from matplotlib import pyplot as plt, axes
 
 from procastro.astrofile import static_identify, static_read, static_guess, static_write
 from procastro.cache.cache import astrofile_cache
-from procastro.cache.cachev2 import astrofile_cachev2
+from procastro.cache.cache import astrofile_cachev2
 from procastro.astrofile.meta import CaseInsensitiveMeta
 from procastro.interfaces import IAstroCalib, IAstroFile
 from procastro.logging import io_logger
